Add AddTimed() to Warnings for timestamped warnings
