+++ RipCheck/Tracks/DrumsTrack/DrumsTrack.cs
@@ -30,13 +30,7 @@ namespace RipCheck
 
                 if (!Enum.IsDefined(typeof(DrumsTrackNotes), key))
                 {
-                    var position = note.Time;
-                    var time = (MetricTimeSpan) TimeConverter.ConvertTo(position, TimeSpanType.Metric, tempoMap);
-                    var ticks = (BarBeatTicksTimeSpan) TimeConverter.ConvertTo(position, TimeSpanType.BarBeatTicks, tempoMap);
-                    int minutes = 60 * time.Hours + time.Minutes;
-                    int seconds = time.Seconds;
-                    int millisecs = time.Milliseconds;
-                    trackWarnings.Add($"Unknown note: {key} on {name} at {minutes}:{seconds:d2}.{millisecs:d3} (MBT: {ticks.Bars}.{ticks.Beats}.{ticks.Ticks})");
+                    trackWarnings.AddTimed($"Unknown note: {key} on {name}", note.Time, tempoMap);
                     continue;
                 }
 
@@ -70,12 +64,7 @@ namespace RipCheck
                 {
                     if (gap > 0 && gap < 10)
                     {
-                        var time = (MetricTimeSpan) TimeConverter.ConvertTo(position, TimeSpanType.Metric, tempoMap);
-                        var ticks = (BarBeatTicksTimeSpan) TimeConverter.ConvertTo(position, TimeSpanType.BarBeatTicks, tempoMap);
-                        int minutes = 60 * time.Hours + time.Minutes;
-                        int seconds = time.Seconds;
-                        int millisecs = time.Milliseconds;
-                        warnings.Add($"Chord snapping: {name} {difficulty} at {minutes}:{seconds:d2}.{millisecs:d3} (MBT: {ticks.Bars}.{ticks.Beats}.{ticks.Ticks})");
+                        trackWarnings.AddTimed($"Chord snapping: {difficulty} on {name}", position, tempoMap);
                     }
                 }
             }
