+++ RipCheck/Tracks/ProGuitarTrack/ProGuitarTrack.cs
@@ -31,13 +31,7 @@ namespace RipCheck
 
                 if (!Enum.IsDefined(typeof(ProGuitarTrackNote), key))
                 {
-                    var position = note.Time;
-                    var time = (MetricTimeSpan) TimeConverter.ConvertTo(position, TimeSpanType.Metric, tempoMap);
-                    var ticks = (BarBeatTicksTimeSpan) TimeConverter.ConvertTo(position, TimeSpanType.BarBeatTicks, tempoMap);
-                    int minutes = 60 * time.Hours + time.Minutes;
-                    int seconds = time.Seconds;
-                    int millisecs = time.Milliseconds;
-                    trackWarnings.Add($"Unknown note: {key} on {name} at {minutes}:{seconds:d2}.{millisecs:d3} (MBT: {ticks.Bars}.{ticks.Beats}.{ticks.Ticks})");
+                    trackWarnings.AddTimed($"Unknown note: {key} on {name}", note.Time, tempoMap);
                     continue;
                 }
 
@@ -48,13 +42,8 @@ namespace RipCheck
 
                 if (velocity < 100 || velocity > 122)
                 {
-                    var position = note.Time;
-                    var time = (MetricTimeSpan) TimeConverter.ConvertTo(position, TimeSpanType.Metric, tempoMap);
-                    var ticks = (BarBeatTicksTimeSpan) TimeConverter.ConvertTo(position, TimeSpanType.BarBeatTicks, tempoMap);
-                    int minutes = 60 * time.Hours + time.Minutes;
-                    int seconds = time.Seconds;
-                    int millisecs = time.Milliseconds;
-                    trackWarnings.Add($"Invalid Pro Guitar fret number: note {key} with velocity {velocity} on {name} at {minutes}:{seconds:d2}.{millisecs:d3} (MBT: {ticks.Bars}.{ticks.Beats}.{ticks.Ticks})");
+                    trackWarnings.AddTimed($"Invalid fret number: note {key} with velocity {velocity} on {name}", note.Time, tempoMap);
+                    continue;
                 }
 
                 Difficulty difficulty = (Difficulty)((key - 24) / 24);
@@ -84,12 +73,7 @@ namespace RipCheck
                 {
                     if (gap > 0 && gap < 10)
                     {
-                        var time = (MetricTimeSpan) TimeConverter.ConvertTo(position, TimeSpanType.Metric, tempoMap);
-                        var ticks = (BarBeatTicksTimeSpan) TimeConverter.ConvertTo(position, TimeSpanType.BarBeatTicks, tempoMap);
-                        int minutes = 60 * time.Hours + time.Minutes;
-                        int seconds = time.Seconds;
-                        int millisecs = time.Milliseconds;
-                        warnings.Add($"Chord snapping: {name} {difficulty} at {minutes}:{seconds:d2}.{millisecs:d3} (MBT: {ticks.Bars}.{ticks.Beats}.{ticks.Ticks})");
+                        trackWarnings.AddTimed($"Chord snapping: {difficulty} on {name}", position, tempoMap);
                     }
                 }
             }
@@ -120,12 +104,7 @@ namespace RipCheck
                     {
                         continue;
                     }
-                    var lateTime = (MetricTimeSpan) TimeConverter.ConvertTo(latePos, TimeSpanType.Metric, tempoMap);
-                    var ticks = (BarBeatTicksTimeSpan) TimeConverter.ConvertTo(latePos, TimeSpanType.BarBeatTicks, tempoMap);
-                    int minutes = 60 * lateTime.Hours + lateTime.Minutes;
-                    int seconds = lateTime.Seconds;
-                    int millisecs = lateTime.Milliseconds;
-                    warnings.Add($"Disjoint chord: {name} {difficulty} at {minutes}:{seconds:d2}.{millisecs:d3} (MBT: {ticks.Bars}.{ticks.Beats}.{ticks.Ticks})");
+                    trackWarnings.AddTimed($"Disjoint chord: {difficulty} on {name}", latePos, tempoMap);
                 }
             }
